Add expand/collapse all links
Support for elements which toggle between expanding and collapsing all accordion panels in their section. Elements should be placed inside of the js-accordion div and given a class of js-accordion__toggle. There is an optional data-collapse-text attribute for the toggle text and "collapse all" will be used if no text is given.

Example of a toggle link:
<a href="#" class="js-accordion__toggle" data-collapse-text="(collapse all)">(expand all)</a>